GChip Refactor, add demo for chips with gavatar/gicon
